major simplificaiton of json writing for the scattering matrix; move linewidth writing to vectorBTE object

+++ src/bte/general_scattering.cpp
@@ -40,7 +40,7 @@ void addBoundaryScattering(ScatteringMatrix &matrix, Context &context,
   #pragma omp parallel for default(none) shared(                            \
    bandStructure, numCalculations, statisticsSweep, boundaryLength,   \
    particle, outPopulations, inPopulations, linewidth, switchCase, excludeIndices, is1s, matrix)
-  for (int iis1 = 0; iis1 < is1s.size(); iis1++ ) {
+  for (size_t iis1 = 0; iis1 < size_t(is1s.size()); iis1++ ) {
 
       int is1 = is1s[iis1]; // we do this because an iterator cannot be omp parallelized
